hotel_booking/requirements.txt
# argon2-cffi==23.1.0
# argon2-cffi-bindings==21.2.0
# asgiref==3.7.2
# asn1crypto==1.5.1
# backcall==0.2.0
# certifi==2024.6.2
# cffi==1.16.0
# chardet==5.2.0
# charset-normalizer==3.3.2
# confusable-homoglyphs==3.3.1

cryptography==42.0.8
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.2
django-allauth==0.63.3
django-configurations==2.5.1

djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jeepney==0.8.0
jwt==1.3.1
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
pipfile==0.0.2

pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python-magic==0.4.27
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
SecretStorage==3.3.3
setuptools==70.0.0
simplegeneric==0.8.1
six==1.16.0
sqlparse==0.4.4
ssh-import-id==5.11
toml==0.10.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.43.0
